refactor: rename travel folder to ride
