+++ app/(tabs)/main/estimate.tsx
@@ -7,7 +7,7 @@ import Input from "@/components/input/Input"
 import EditRideDirectionModal from "@/components/modal/rideModal/EditRideDirectionModal"
 import EditRideRouteModal from "@/components/modal/rideModal/EditRideRouteModal"
 import EditRideStopModal from "@/components/modal/rideModal/EditRideStopModal"
-import { JustifiedLabelValue } from "@/components/travel/IndividualTravelDetailCard"
+import { JustifiedLabelValue } from "@/components/ride/IndividualTravelDetailCard"
 import { useTheme } from "@/context/ThemeContext"
 import useDirections from "@/hooks/data/useDirections"
 import useRoutes from "@/hooks/data/useRoutes"
