+++ app/(tabs)/main/index.tsx
@@ -4,7 +4,7 @@ import CustomIcon from "@/components/CustomIcon"
 import Input from "@/components/input/Input"
 import LoadingScreen from "@/components/LoadingScreen"
 import CalendarModal from "@/components/modal/CalendarModal"
-import GroupedDataDisplay from "@/components/travel/GroupedTravelsDisplay"
+import GroupedDataDisplay from "@/components/ride/GroupedTravelsDisplay"
 import { useSettings } from "@/context/SettingsContext"
 import { useTheme } from "@/context/ThemeContext"
 import useLaps from "@/hooks/data/useLaps"
