+++ app/(tabs)/main/travelDetail.tsx
@@ -5,7 +5,7 @@ import Container from '@/components/Container'
 import Input from '@/components/input/Input'
 import LoadingScreen from '@/components/LoadingScreen'
 import MapDisplay from '@/components/MapDisplay'
-import IndividualTravelDetailCard from '@/components/travel/IndividualTravelDetailCard'
+import IndividualTravelDetailCard from '@/components/ride/IndividualTravelDetailCard'
 import { useTheme } from '@/context/ThemeContext'
 import { useTravelContext } from '@/context/TravelContext'
 import useLaps from '@/hooks/data/useLaps'
