+++ app/(tabs)/manage/datalist.tsx
@@ -6,7 +6,7 @@ import Input from '@/components/input/Input'
 import { TextInputBase } from '@/components/input/TextInput'
 import LoadingScreen from '@/components/LoadingScreen'
 import ModalTemplate from '@/components/ModalTemplate'
-import { EmptyHeaderComponent } from '@/components/travel/TravelFlatlist'
+import { EmptyHeaderComponent } from '@/components/ride/TravelFlatlist'
 import { useDataEditContext } from '@/context/DataEditContext'
 import { useDialog } from '@/context/DialogContext'
 import useDirections from '@/hooks/data/useDirections'
