+++ components/ride/IndividualTravelDetailCard.tsx
